plugin-geogebra: migrate to EditorPlugin
