+++ packages/private/bundle-size/src/shared/plugins.ts
@@ -8,7 +8,7 @@ import {
   parseFileType,
   UploadedFile
 } from '@edtr-io/plugin-files'
-import { geogebraPlugin } from '@edtr-io/plugin-geogebra'
+import { createGeogebraPlugin } from '@edtr-io/plugin-geogebra'
 import { highlightPlugin } from '@edtr-io/plugin-highlight'
 import { hintPlugin } from '@edtr-io/plugin-hint'
 import { inputExercisePlugin } from '@edtr-io/plugin-input-exercise'
@@ -40,7 +40,7 @@ export const plugins: Record<
   blockquote: createBlockquotePlugin(),
   equations: createEquationsPlugin(),
   files: createFilesPlugin({ upload: mockUploadFileHandler }),
-  geogebra: geogebraPlugin,
+  geogebra: createGeogebraPlugin(),
   highlight: highlightPlugin,
   hint: hintPlugin,
   inputExercise: inputExercisePlugin,
